fls_backend/requirements.txt
djangorestframework
Django~=3.0.5
knox
Jinja2
django-enumchoicefield